Sort stocktake insert into respective folder/files
