+++ server/service/src/stocktake/insert/mod.rs
@@ -1,17 +1,15 @@
-use chrono::{NaiveDate, Utc};
-use repository::{
-    ActivityLogType, DateFilter, EqualFilter, ItemRowRepository, ItemType, MasterListFilter,
-    MasterListLineFilter, MasterListLineRepository, MasterListRepository, NumberRowType,
-    RepositoryError, StockLineFilter, StockLineRepository, StockLineRow, Stocktake,
-    StocktakeFilter, StocktakeLineRow, StocktakeLineRowRepository, StocktakeRepository,
-    StocktakeRow, StocktakeRowRepository, StocktakeStatus, StorageConnection,
-};
-use util::uuid::uuid;
+mod validate;
+use validate::validate;
 
-use crate::{
-    activity_log::activity_log_entry, check_location_exists, number::next_number,
-    service_provider::ServiceContext, validate::check_store_exists, NullableUpdate,
+mod generate;
+use generate::generate;
+
+use chrono::NaiveDate;
+use repository::{
+    ActivityLogType, RepositoryError, Stocktake, StocktakeLineRowRepository, StocktakeRowRepository,
 };
+
+use crate::{activity_log::activity_log_entry, service_provider::ServiceContext, NullableUpdate};
 
 use super::query::get_stocktake;
 
@@ -37,400 +35,6 @@ pub enum InsertStocktakeError {
     InvalidMasterList,
     InvalidLocation,
     InvalidArguments,
-}
-
-fn check_stocktake_does_not_exist(
-    connection: &StorageConnection,
-    id: &str,
-) -> Result<bool, RepositoryError> {
-    let count = StocktakeRepository::new(connection)
-        .count(Some(StocktakeFilter::new().id(EqualFilter::equal_to(id))))?;
-    Ok(count == 0)
-}
-
-fn check_master_list_exists(
-    connection: &StorageConnection,
-    store_id: &str,
-    master_list_id: &str,
-) -> Result<bool, RepositoryError> {
-    let count = MasterListRepository::new(connection).count(Some(
-        MasterListFilter::new()
-            .id(EqualFilter::equal_to(master_list_id))
-            .exists_for_store_id(EqualFilter::equal_to(store_id)),
-    ))?;
-    Ok(count > 0)
-}
-
-fn validate(
-    connection: &StorageConnection,
-    store_id: &str,
-    stocktake: &InsertStocktake,
-) -> Result<(), InsertStocktakeError> {
-    if !check_stocktake_does_not_exist(connection, &stocktake.id)? {
-        return Err(InsertStocktakeError::StocktakeAlreadyExists);
-    }
-    if !check_store_exists(connection, store_id)? {
-        return Err(InsertStocktakeError::InvalidStore);
-    }
-    if stocktake.master_list_id.is_some() && stocktake.location.is_some() {
-        return Err(InsertStocktakeError::InvalidArguments);
-    }
-    if let Some(master_list_id) = &stocktake.master_list_id {
-        if !check_master_list_exists(connection, store_id, master_list_id)? {
-            return Err(InsertStocktakeError::InvalidMasterList);
-        }
-    }
-
-    if !check_location_exists(connection, store_id, &stocktake.location)? {
-        return Err(InsertStocktakeError::InvalidLocation);
-    }
-
-    Ok(())
-}
-
-fn generate(
-    connection: &StorageConnection,
-    store_id: &str,
-    user_id: &str,
-    InsertStocktake {
-        id,
-        comment,
-        description,
-        stocktake_date,
-        is_locked,
-        location,
-        master_list_id,
-        items_have_stock,
-        expires_before,
-    }: InsertStocktake,
-) -> Result<(StocktakeRow, Vec<StocktakeLineRow>), RepositoryError> {
-    let stocktake_number = next_number(connection, &NumberRowType::Stocktake, store_id)?;
-
-    let master_list_lines = match master_list_id {
-        Some(master_list_id) => {
-            generate_lines_from_master_list(connection, store_id, &id, &master_list_id)?
-        }
-        None => Vec::new(),
-    };
-    let location_lines = match location {
-        Some(NullableUpdate {
-            value: Some(location_id),
-            ..
-        }) => generate_lines_from_location(connection, store_id, &id, &location_id)?,
-        _ => Vec::new(),
-    };
-    let items_have_stock_lines = match items_have_stock {
-        Some(_) => generate_lines_with_stock(connection, store_id, &id)?,
-        None => Vec::new(),
-    };
-    let expiring_items_lines = match expires_before {
-        Some(expires_before) => {
-            generate_lines_expiring_before(connection, store_id, &id, &expires_before)?
-        }
-        None => Vec::new(),
-    };
-    let lines = [
-        master_list_lines,
-        location_lines,
-        items_have_stock_lines,
-        expiring_items_lines,
-    ]
-    .concat();
-
-    Ok((
-        StocktakeRow {
-            id,
-            stocktake_number,
-            comment,
-            description,
-            // TODO: Changing this to be same as created datetime for now since function is disabled in frontend
-            // but will need to remove this later when functionality is
-            stocktake_date: if stocktake_date.is_some() {
-                stocktake_date
-            } else {
-                Some(Utc::now().naive_utc().date())
-            },
-            status: StocktakeStatus::New,
-            created_datetime: Utc::now().naive_utc(),
-            user_id: user_id.to_string(),
-            store_id: store_id.to_string(),
-            is_locked: is_locked.unwrap_or(false),
-            // Default
-            finalised_datetime: None,
-            inventory_addition_id: None,
-            inventory_reduction_id: None,
-        },
-        lines,
-    ))
-}
-
-fn generate_lines_from_master_list(
-    connection: &StorageConnection,
-    store_id: &str,
-    stocktake_id: &str,
-    master_list_id: &str,
-) -> Result<Vec<StocktakeLineRow>, RepositoryError> {
-    let item_ids: Vec<String> = MasterListLineRepository::new(connection)
-        .query_by_filter(
-            MasterListLineFilter::new()
-                .master_list_id(EqualFilter::equal_to(master_list_id))
-                .item_type(ItemType::Stock.equal_to()),
-        )?
-        .into_iter()
-        .map(|r| r.item_id)
-        .collect();
-
-    let mut result = Vec::<StocktakeLineRow>::new();
-
-    item_ids.iter().for_each(|item_id| {
-        let stock_lines = StockLineRepository::new(connection)
-            .query_by_filter(
-                StockLineFilter::new()
-                    .item_id(EqualFilter::equal_to(item_id))
-                    .store_id(EqualFilter::equal_to(store_id))
-                    .has_packs_in_store(true),
-                Some(store_id.to_string()),
-            )
-            .unwrap();
-        let item_name = ItemRowRepository::new(connection)
-            .find_active_by_id(item_id)
-            .unwrap()
-            .unwrap()
-            .name;
-
-        if stock_lines.is_empty() {
-            result.push(StocktakeLineRow {
-                id: uuid(),
-                stocktake_id: stocktake_id.to_string(),
-                snapshot_number_of_packs: 0.0,
-                item_link_id: item_id.to_string(),
-                item_name,
-                location_id: None,
-                batch: None,
-                expiry_date: None,
-                note: None,
-                stock_line_id: None,
-                pack_size: None,
-                cost_price_per_pack: None,
-                sell_price_per_pack: None,
-                comment: None,
-                counted_number_of_packs: None,
-                inventory_adjustment_reason_id: None,
-            });
-        } else {
-            stock_lines.into_iter().for_each(|line| {
-                let StockLineRow {
-                    id: stock_line_id,
-                    item_link_id: _,
-                    location_id,
-                    batch,
-                    pack_size,
-                    cost_price_per_pack,
-                    sell_price_per_pack,
-                    total_number_of_packs,
-                    expiry_date,
-                    note,
-                    supplier_link_id: _,
-                    store_id: _,
-                    on_hold: _,
-                    available_number_of_packs: _,
-                    barcode_id: _,
-                } = line.stock_line_row;
-
-                result.push(StocktakeLineRow {
-                    id: uuid(),
-                    stocktake_id: stocktake_id.to_string(),
-                    snapshot_number_of_packs: total_number_of_packs,
-                    item_link_id: line.item_row.id,
-                    item_name: line.item_row.name,
-                    location_id,
-                    batch,
-                    expiry_date,
-                    note,
-                    stock_line_id: Some(stock_line_id),
-                    pack_size: Some(pack_size),
-                    cost_price_per_pack: Some(cost_price_per_pack),
-                    sell_price_per_pack: Some(sell_price_per_pack),
-                    comment: None,
-                    counted_number_of_packs: None,
-                    inventory_adjustment_reason_id: None,
-                });
-            });
-        }
-    });
-
-    Ok(result)
-}
-
-fn generate_lines_from_location(
-    connection: &StorageConnection,
-    store_id: &str,
-    stocktake_id: &str,
-    location_id: &str,
-) -> Result<Vec<StocktakeLineRow>, RepositoryError> {
-    let stock_lines = StockLineRepository::new(connection).query_by_filter(
-        StockLineFilter::new()
-            .location_id(EqualFilter::equal_to(location_id))
-            .store_id(EqualFilter::equal_to(store_id))
-            .has_packs_in_store(true),
-        Some(store_id.to_string()),
-    )?;
-
-    let result = stock_lines
-        .into_iter()
-        .map(|line| {
-            let StockLineRow {
-                id: stock_line_id,
-                item_link_id: _,
-                location_id,
-                batch,
-                pack_size,
-                cost_price_per_pack,
-                sell_price_per_pack,
-                total_number_of_packs,
-                expiry_date,
-                note,
-                supplier_link_id: _,
-                store_id: _,
-                on_hold: _,
-                available_number_of_packs: _,
-                barcode_id: _,
-            } = line.stock_line_row;
-
-            StocktakeLineRow {
-                id: uuid(),
-                stocktake_id: stocktake_id.to_string(),
-                snapshot_number_of_packs: total_number_of_packs,
-                item_link_id: line.item_row.id,
-                item_name: line.item_row.name,
-                location_id,
-                batch,
-                expiry_date,
-                note,
-                stock_line_id: Some(stock_line_id),
-                pack_size: Some(pack_size),
-                cost_price_per_pack: Some(cost_price_per_pack),
-                sell_price_per_pack: Some(sell_price_per_pack),
-                comment: None,
-                counted_number_of_packs: None,
-                inventory_adjustment_reason_id: None,
-            }
-        })
-        .collect();
-    Ok(result)
-}
-
-pub fn generate_lines_with_stock(
-    connection: &StorageConnection,
-    store_id: &str,
-    stocktake_id: &str,
-) -> Result<Vec<StocktakeLineRow>, RepositoryError> {
-    let stock_lines = StockLineRepository::new(connection).query_by_filter(
-        StockLineFilter::new()
-            .store_id(EqualFilter::equal_to(store_id))
-            .has_packs_in_store(true),
-        Some(store_id.to_string()),
-    )?;
-
-    let result = stock_lines
-        .into_iter()
-        .map(|line| {
-            let StockLineRow {
-                id: stock_line_id,
-                item_link_id: _,
-                location_id,
-                batch,
-                pack_size,
-                cost_price_per_pack,
-                sell_price_per_pack,
-                total_number_of_packs,
-                expiry_date,
-                note,
-                supplier_link_id: _,
-                store_id: _,
-                on_hold: _,
-                available_number_of_packs: _,
-                barcode_id: _,
-            } = line.stock_line_row;
-
-            StocktakeLineRow {
-                id: uuid(),
-                stocktake_id: stocktake_id.to_string(),
-                snapshot_number_of_packs: total_number_of_packs,
-                item_link_id: line.item_row.id,
-                item_name: line.item_row.name,
-                location_id,
-                batch,
-                expiry_date,
-                note,
-                stock_line_id: Some(stock_line_id),
-                pack_size: Some(pack_size),
-                cost_price_per_pack: Some(cost_price_per_pack),
-                sell_price_per_pack: Some(sell_price_per_pack),
-                comment: None,
-                counted_number_of_packs: None,
-                inventory_adjustment_reason_id: None,
-            }
-        })
-        .collect();
-    Ok(result)
-}
-
-fn generate_lines_expiring_before(
-    connection: &StorageConnection,
-    store_id: &str,
-    stocktake_id: &str,
-    date: &NaiveDate,
-) -> Result<Vec<StocktakeLineRow>, RepositoryError> {
-    let stock_lines = StockLineRepository::new(connection).query_by_filter(
-        StockLineFilter::new()
-            .store_id(EqualFilter::equal_to(store_id))
-            .expiry_date(DateFilter::before_or_equal_to(*date)),
-        Some(store_id.to_string()),
-    )?;
-
-    let result = stock_lines
-        .into_iter()
-        .map(|line| {
-            let StockLineRow {
-                id: stock_line_id,
-                item_link_id: _,
-                location_id,
-                batch,
-                pack_size,
-                cost_price_per_pack,
-                sell_price_per_pack,
-                total_number_of_packs,
-                expiry_date,
-                note,
-                supplier_link_id: _,
-                store_id: _,
-                on_hold: _,
-                available_number_of_packs: _,
-                barcode_id: _,
-            } = line.stock_line_row;
-
-            StocktakeLineRow {
-                id: uuid(),
-                stocktake_id: stocktake_id.to_string(),
-                snapshot_number_of_packs: total_number_of_packs,
-                item_link_id: line.item_row.id,
-                location_id,
-                batch,
-                expiry_date,
-                note,
-                stock_line_id: Some(stock_line_id),
-                pack_size: Some(pack_size),
-                cost_price_per_pack: Some(cost_price_per_pack),
-                sell_price_per_pack: Some(sell_price_per_pack),
-                comment: None,
-                counted_number_of_packs: None,
-                inventory_adjustment_reason_id: None,
-                item_name: line.item_row.name,
-            }
-        })
-        .collect();
-    Ok(result)
 }
 
 pub fn insert_stocktake(
