+++ server/service/src/stocktake/mod.rs
@@ -3,7 +3,6 @@ use repository::PaginationOption;
 use repository::{RepositoryError, Stocktake, StocktakeFilter, StocktakeSort};
 
 pub mod query;
-pub mod validate;
 
 mod delete;
 pub use self::delete::*;
@@ -17,6 +16,9 @@ pub use self::update::*;
 
 mod batch;
 pub use self::batch::*;
+
+mod common;
+pub use self::common::*;
 
 pub trait StocktakeServiceTrait: Sync + Send {
     fn get_stocktakes(
